Move UnnecessaryCodeFormatDefinition to EditorFeatures.Wpf
ClassificationFormatDefinition are only defined in
Microsoft.VisualStudio.Text.UI.Wpf.
